Optimize _check_auth_results and _check_missing_headers
